fix: 32bit systems support №2
change other usages to int64

+++ mtproto.go
@@ -453,7 +453,7 @@ func (m *MTProto) readMsg() error {
 	response, err := m.transport.ReadMsg()
 	if err != nil {
 		if e, ok := err.(transport.ErrCode); ok {
-			return &ErrResponseCode{Code: int(e)}
+			return &ErrResponseCode{Code: int64(e)}
 		}
 		switch err {
 		case io.EOF, context.Canceled:
